+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.0.7 (2019-06-18)
+* Dropped support for Python 2
+* Supports Google docstring format
+
 ## 0.0.6 (2018-12-28)
 * Python3 support. Finally!